+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @queryweave/nuxt
 
+## 0.1.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @queryweave/core@0.1.0-beta.4
+  - @queryweave/vue@0.1.0-beta.4
+  - @queryweave/vue-router@0.1.0-beta.4
+
 ## 0.1.0-beta.3
 
 ### Patch Changes